# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "summary": "Get a category by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the category.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found."
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/admin/categories": {
      "get": {
        "operationId": "AdminCategoriesController_findAll",
        "summary": "List all categories (admin view)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all categories with admin fields.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminCategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AdminCategoriesController_create",
        "summary": "Create a new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/categories/{id}": {
      "get": {
        "operationId": "AdminCategoriesController_findOne",
        "summary": "Get a category by id (admin view)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the category with admin fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found."
          }
        },
        "tags": [
          "Admin Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AdminCategoriesController_update",
        "summary": "Update a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found."
          }
        },
        "tags": [
          "Admin Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_remove",
        "summary": "Delete a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The category has been successfully deleted."
          },
          "404": {
            "description": "Category not found."
          }
        },
        "tags": [
          "Admin Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "The Sync App API",
    "description": "This a the project document",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "X-AI-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AI-API-Key",
        "description": "AI Service API Token"
      }
    },
    "schemas": {
...
      "CategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique category identifier"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug"
          },
          "description": {
            "type": "object",
            "description": "Category description"
          },
          "imageUrl": {
            "type": "object",
            "description": "Category image URL"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether category is active"
          },
          "categoryType": {
            "type": "string",
            "description": "Category type for UI grouping",
            "default": "primary"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          },
          "parent": {
            "description": "Parent category",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategorySummaryDto"
              }
            ]
          },
          "children": {
            "description": "Child categories",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummaryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AdminCategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique category identifier"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug"
          },
          "description": {
            "type": "object",
            "description": "Category description"
          },
          "imageUrl": {
            "type": "object",
            "description": "Category image URL"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether category is active"
          },
          "iconName": {
            "type": "object",
            "description": "Icon identifier for frontend rendering"
          },
          "colorValue": {
            "type": "object",
            "description": "Hex color value (e.g. #FF6B6B)"
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order for display (lower = first)"
          },
          "serviceCount": {
            "type": "number",
            "description": "Number of health services in this category"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          },
          "parent": {
            "description": "Parent category",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategorySummaryDto"
              }
            ]
          },
          "children": {
            "description": "Child categories",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummaryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "isActive",
          "sortOrder",
          "serviceCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "string",
            "example": "uuid-parent-id",
            "description": "Parent category ID for hierarchical structure"
          },
          "name": {
            "type": "string",
            "example": "Massage Services",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "example": "massage-services",
            "description": "URL-friendly slug (unique)"
          },
          "description": {
            "type": "string",
            "example": "All massage related services",
            "description": "Category description"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/category.jpg",
            "description": "Category image URL"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the category is active"
          },
          "iconName": {
            "type": "string",
            "example": "spa",
            "description": "Icon identifier for frontend rendering"
          },
          "colorValue": {
            "type": "string",
            "example": "#FF6B6B",
            "description": "Hex color value (e.g. #FF6B6B or #FF6B6BCC)"
          },
          "sortOrder": {
            "type": "number",
            "example": 0,
            "description": "Sort order for display (lower = first)"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "string",
            "example": "uuid-parent-id",
            "description": "Parent category ID for hierarchical structure"
          },
          "name": {
            "type": "string",
            "example": "Massage Services",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "example": "massage-services",
            "description": "URL-friendly slug (unique)"
          },
          "description": {
            "type": "string",
            "example": "All massage related services",
            "description": "Category description"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/category.jpg",
            "description": "Category image URL"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the category is active"
          },
          "iconName": {
            "type": "string",
            "example": "spa",
            "description": "Icon identifier for frontend rendering"
          },
          "colorValue": {
            "type": "string",
            "example": "#FF6B6B",
            "description": "Hex color value (e.g. #FF6B6B or #FF6B6BCC)"
          },
          "sortOrder": {
            "type": "number",
            "example": 0,
            "description": "Sort order for display (lower = first)"
          }
        }
      },
...
    }
  }
}